# nuxk-core/api/openapi.yaml
openapi: 3.1.0
info:
  title: nuxk-core agent API
  version: "1"
  description: >
    The contract between the nuxk-core agent (on the router, or the Pi test
    stand) and its clients: the lite web UI served by the agent itself and
    nuxk-controller on the Pi.

    Single-tenant. Every path except /healthz and /auth/* needs either
    `Authorization: Bearer <API_TOKEN>` (`/opt/etc/nuxk/nuxk.conf`; programs —
    the controller) or the `nuxk_agent` session cookie from POST /auth/login
    (a browser: the box's root login and password, as for SSH). A session's
    writes must come from the agent's own page (Origin is checked). A loopback
    caller is let in without either only when no token is configured.
    Responses are the resource itself; errors are `{"error":{"code","message"}}`
    plus the HTTP status. Request bodies are capped at 4 MiB.

    This file is checked against the agent's route table in CI
    (TestOpenAPIMatchesRoutes): an endpoint exists here if and only if the
    agent serves it. Planned endpoints are not listed until they exist.
  license: { name: MIT }
servers:
  - url: http://{host}:4141
    variables: { host: { default: 192.168.1.1 } }
security:
  - bearer: []
  - session: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Log a browser in with the box's root account
      description: >
        Checks the password against Entware's /opt/etc/shadow (AUTH_USER /
        AUTH_FILES in nuxk.conf) and sets the `nuxk_agent` session cookie
        (HttpOnly, SameSite=Strict; forgotten after 12 h unused or a restart
        of nuxk-core). After 5 failures from one address the next attempts
        wait, doubling up to 15 min.
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Credentials" } } }
      responses:
        "200":
          description: logged in; Set-Cookie carries the session
          content: { application/json: { schema: { $ref: "#/components/schemas/SessionInfo" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
...
  /api/v1/auth/pair:
    post:
      tags: [auth]
      operationId: pair
      summary: Hand the API token to a controller that knows the root password
      description: >
        nuxk-controller's setup calls this once with the root login and
        password the person typed; it keeps the token, never the password.
        Same lockout as /auth/login.
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/Credentials" } } }
      responses:
        "200":
          description: the agent's API_TOKEN
          content: { application/json: { schema: { $ref: "#/components/schemas/PairInfo" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
...
  /api/v1/info:
    get:
      tags: [agent]
      operationId: info
      summary: What box this is (router or test stand), firmware, uptime
      responses:
        "200":
          description: node
          content: { application/json: { schema: { $ref: "#/components/schemas/NodeInfo" } } }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/status:
    get:
      tags: [agent]
      operationId: status
      summary: Whole platform state — every engine plus the routing plane
      description: Served from memory (the reconcile loop keeps it fresh); never blocks on an engine.
      responses:
        "200":
          description: snapshot
          content: { application/json: { schema: { $ref: "#/components/schemas/Status" } } }
        "401": { $ref: "#/components/responses/Error" }
...
  /api/v1/logs:
    get:
      tags: [agent]
      operationId: logs
      summary: The agent's recent log (last 2000 entries kept in memory)
      parameters:
        - { name: after, in: query, description: only entries with a larger seq, schema: { type: integer, minimum: 0 } }
        - { name: limit, in: query, description: "at most this many, newest (default 200, max 2000)", schema: { type: integer, minimum: 1, maximum: 2000 } }
      responses:
        "200":
          description: entries, oldest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/LogEntry" } }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/logs/debug:
    get:
      tags: [agent]
      operationId: logsDebug
      summary: The debug switch of the in-memory log
      responses:
        "200":
          description: the switch
          content: { application/json: { schema: { $ref: "#/components/schemas/LogDebug" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [agent]
      operationId: setLogsDebug
      summary: Debug logging on for a while (it turns itself off), or off
      description: >
        While on, debug entries — engine script calls, RCI requests, commands,
        DNS failures — reach the in-memory log (/logs, /events). The log file on
        the router's flash keeps info and up. The agent started with -debug is
        on for good ("forced").
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LogDebugSet" } } }
      responses:
        "200":
          description: the switch after the change
          content: { application/json: { schema: { $ref: "#/components/schemas/LogDebug" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/events:
    get:
      tags: [agent]
      operationId: events
      summary: Live stream (Server-Sent Events)
      description: >
        `event: status` with a Status snapshot on connect and on every change;
        `event: log` with a LogEntry per new log line; a `: ping` comment every
        20 s. The stream stays open until the client goes away.
      responses:
        "200":
          description: text/event-stream
          content: { text/event-stream: { schema: { type: string } } }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/engines:
    get:
      tags: [engines]
      operationId: engines
      responses:
        "200":
          description: every wired engine
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/EngineState" } }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/engines/{kind}:
    get:
      tags: [engines]
      operationId: engine
      summary: One engine, read live (not from the snapshot)
      parameters: [{ $ref: "#/components/parameters/kind" }]
      responses:
        "200":
          description: live info
          content: { application/json: { schema: { $ref: "#/components/schemas/EngineInfo" } } }
        "404": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/engines/{kind}/{action}:
    post:
      tags: [engines]
      operationId: engineAction
      summary: start | stop | restart | probe | apply
      description: >
        start/stop/restart also record the run intent (the controller keeps it
        with backoff). probe runs an active check now (nfqws2: its probe sites,
        see probe-targets), keeps it as the engine's last probe and returns it.
        apply takes a Routing body.
      parameters:
        - { $ref: "#/components/parameters/kind" }
        - { name: action, in: path, required: true, schema: { type: string, enum: [start, stop, restart, probe, apply] } }
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: "#/components/schemas/Routing" } } }
      responses:
        "200":
          description: probe result
          content: { application/json: { schema: { $ref: "#/components/schemas/Probe" } } }
        "202":
          description: done
          content: { application/json: { schema: { $ref: "#/components/schemas/Ok" } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/engines/{kind}/config:
    put:
      tags: [engines]
      operationId: engineConfig
      summary: Runtime config (xray only today); stored 0600, never served back
      description: >
        xray: exactly one of vless_uri (a vless:// link) or sub_url (a 3x-ui
        subscription; its first VLESS server). The agent renders config.json
        (TUN inbound opkgtunN → VLESS), xray tests it, and it replaces the old
        one; if xray won't start with it, the old one comes back (502). A link
        that can't work is 400 with the reason in words.
      parameters: [{ $ref: "#/components/parameters/kind" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/EngineConfig" } } }
      responses:
        "200":
          description: applied; what the config now points at
          content: { application/json: { schema: { $ref: "#/components/schemas/Upstream" } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/engines/{kind}/upstream:
    get:
      tags: [engines]
      operationId: engineUpstream
      summary: What the engine's config points at (xray) — no secrets
      description: >
        The servers of the link or subscription, the one in use, and what the
        subscription's panel reports (3x-ui: title, traffic, term, how often
        to re-read it). A subscription is re-read on the router every
        Profile-Update-Interval (12 h if unset, within 1–24 h); the same
        server with the same settings restarts nothing.
      parameters: [{ $ref: "#/components/parameters/kind" }]
      responses:
        "200":
          description: source "" = nothing set yet
          content: { application/json: { schema: { $ref: "#/components/schemas/Upstream" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/engines/{kind}/upstream/pick:
    post:
      tags: [engines]
      operationId: engineUpstreamPick
      summary: Switch to another server of the stored subscription
      parameters: [{ $ref: "#/components/parameters/kind" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [server]
              properties:
                server: { type: string, description: "an UpstreamServer key (name@host:port)" }
      responses:
        "200":
          description: applied
          content: { application/json: { schema: { $ref: "#/components/schemas/Upstream" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { description: "the server is a link, not a subscription", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/engines/{kind}/upstream/refresh:
    post:
      tags: [engines]
      operationId: engineUpstreamRefresh
      summary: Re-read the stored subscription now
      description: A failure keeps the running config; its reason is in the stored Upstream's error.
      parameters: [{ $ref: "#/components/parameters/kind" }]
      responses:
        "200":
          description: re-read
          content: { application/json: { schema: { $ref: "#/components/schemas/Upstream" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { description: "nothing set yet", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "502": { $ref: "#/components/responses/Error" }
  /api/v1/engines/{kind}/strategies:
    get:
      tags: [engines]
      operationId: engineStrategies
      summary: nuxk's per-domain strategies on this engine (nfqws2 only)
      parameters: [{ $ref: "#/components/parameters/kind" }]
      responses:
        "200":
          description: the stored set
          content: { application/json: { schema: { $ref: "#/components/schemas/StrategySet" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [engines]
      operationId: setEngineStrategies
      summary: Replace nuxk's strategies (an empty list removes them)
      description: >
        nfqws2: each strategy becomes its own profile in NFQWS_ARGS_CUSTOM
        (the stock script puts it before the others) with a hostlist
        nuxk-sN.list of its domains and a protocol filter. Only --payload,
        --lua-desync, --out-range and --in-range are accepted, without files;
        a leading "nfqws2" and --filter-*/--hostlist* are dropped. nfqws2
        restarts; the conf is copied to nfqws2.conf.nuxk-bak first (and the
        original once to .nuxk-orig). If nfqws2 won't start, the old conf is
        restored and the answer is 502. Custom args a person wrote there
        themselves are never overwritten (502 with the reason).
      parameters: [{ $ref: "#/components/parameters/kind" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/StrategySet" } } }
      responses:
        "200":
          description: applied, nfqws2 running with them
          content: { application/json: { schema: { $ref: "#/components/schemas/Ok" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
...
  /api/v1/plane:
    get:
      tags: [plane]
      operationId: plane
      responses:
        "200":
          description: plane status
          content: { application/json: { schema: { $ref: "#/components/schemas/PlaneStatus" } } }
        "404": { description: "plane is off (PLANE= in nuxk.conf)", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/plane/lists:
    get:
      tags: [plane]
      operationId: planeLists
      responses:
        "200":
          description: desired lists
          content: { application/json: { schema: { $ref: "#/components/schemas/PlaneDesired" } } }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [plane]
      operationId: setPlaneLists
      summary: Replace every list and the tunnel-down policy; the plane reconciles right away
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PlaneDesired" } } }
      responses:
        "200":
          description: stored
          content: { application/json: { schema: { $ref: "#/components/schemas/Ok" } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    session: { type: apiKey, in: cookie, name: nuxk_agent }
...
  responses:
    Error:
      description: error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Credentials:
      type: object
      required: [user, password]
      properties:
        user: { type: string, description: "root (AUTH_USER)" }
        password: { type: string, format: password }
    SessionInfo:
      type: object
      required: [user]
      properties:
        user: { type: string }
    PairInfo:
      type: object
      required: [token]
      properties:
        token: { type: string, description: the agent's API_TOKEN }
...